tinylingo/python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
rembg==2.0.50
Pillow==10.1.0
python-multipart==0.0.6